requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
timm>=0.9.0
zarr>=2.14.0
numcodecs>=0.11.0
numpy>=1.24.0
pyyaml>=6.0
